Added a bit of a hack to fix some scrolling issues
When there are many items in the RecyclerView, we can't use an offset to determine the correct RecyclerPosition, because it's too expensive with so many children.

However, if we don't use an offset, when there are few children in the RecyclerView, the RecyclerView doesn't get positioned correctly according to the FastScroll position.

This is kind of a happy medium, and most certainly a hack.