Allow varying item view heights
By default, FastScrollRecyclerView will still assume that each row has the same height (as it did before). When this is not the case, an adapter can implement the new MeasurableAdapter interface, which lets FastScrollRecyclerView figure out the actual heights of views. Implementors will usually implement this method by returning the product of their row's height by the viewCount parameter.