Use Canvas.save() to build against API 28
`Canvas.save(int)` was deprecated in API 26 and removed in API 28, so trying to bulid an app against API 28 with this library results in a ProGuard warning that fails the build. See also #88.

Since `Canvas.save()` is equivalent to `Canvas.save(Canvas.MATRIX_SAVE_FLAG | Canvas.CLIP_SAVE_FLAG)` (see their implementation), and the code is not modifying canvas clip, this replacement is totally equivalent.